Removed no longer needed stuff

+++ ci/regenerate_build.sh
@@ -30,8 +30,10 @@ if [ ${PHP_MAJOR_VERSION} -eq 7 ]; then
 	# Workaround to clean ZE3 kernel from ZE2
 	OUTDATED_KERNEL_FILES=`git status --short | grep ' D ' | awk -F' D ' '{print $2}'`
 
-	echo -e "Going to remove from git index:\n${OUTDATED_KERNEL_FILES}"
-	echo $OUTDATED_KERNEL_FILES | xargs git rm -rf
+	if [ ! -z ${OUTDATED_KERNEL_FILES+x} ]; then
+		echo -e "Going to remove from git index:\n${OUTDATED_KERNEL_FILES}"
+		echo $OUTDATED_KERNEL_FILES | xargs git rm -rf
+	fi
 else
 	echo -e "zephir generate ${ZEND_BACKEND}"
 	zephir generate ${ZEND_BACKEND}
